log-message-processor/requirements.txt
redis==2.10.6
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-zipkin==1.25.0